requirements-salt.txt
Django==1.8.5
pytz==2015.6
django-sekizai==0.8.2
djangorestframework==3.2.4
Markdown==2.6.2
uWSGI==2.0.11.2
psycopg2==2.6.1
